changed the default file names from 'module-n' to more relevant names
