+++ src/scripts/cards.js
@@ -1,0 +1,2 @@
+// BUILD the structure to display data objects on the DOM.
+// Fridgify reference - food.js
