+++ src/scripts/eventListener.js
@@ -1,0 +1,1 @@
+// All event listeners
